# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CPI Artifacts API
  version: 1.0.0
  description: API Telegram Mini App для сбора материалов мероприятий.
servers:
  - url: https://artifacts.62-113-105-225.sslip.io
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/admin/exports:
    get:
      tags: [admin]
      summary: Задания экспорта
      responses:
        '200': { description: Список заданий }
    post:
      tags: [admin]
      summary: Запустить CSV, XLSX или ZIP в фоне
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, kind]
              properties:
                eventId: { type: string, format: uuid }
                kind: { enum: [csv, xlsx, zip] }
      responses:
        '202': { description: Задание принято }
  /api/v1/admin/exports/{exportId}/download:
    get:
      tags: [admin]
      summary: Временный URL готовой выгрузки
      parameters:
        - { name: exportId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: URL выгрузки }
        '409': { description: Выгрузка не готова }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cpi_artifacts_session
  parameters:
    CsrfToken:
      name: x-csrf-token
      in: header
      required: true
      schema: { type: string }
...